[nrf fromtree] Build time optimization (#37433)
* Move some classes out of af-types.h

Move AttributesChangedListener and MarkAttributeDirty out of af-types.h


* Remove dependency to cluster-objects.h


* Remove cluster-objects from chip_data_model.cmake


* [Chef] Fix includes

Fix missing includes after changes in Accessors.h


* [Infineon] Fix includes

Fix missing includes after changes in Accessors.h


* [Silabs] Fix includes

Fix missing includes after changes in Accessors.h


---------

(cherry picked from commit 38cc24bd33f738ef02e2a81d67046325098d2349)
