+++ examples/chef/common/chef-dishwasher-mode-delegate-impl.cpp
@@ -17,6 +17,7 @@
  */
 #include <app-common/zap-generated/attributes/Accessors.h>
 #include <app/util/config.h>
+#include <chef-dishwasher-mode-delegate-impl.h>
 
 using namespace chip;
 using namespace chip::app;
