+++ examples/lighting-app/infineon/cyw30739/src/ZclCallbacks.cpp
@@ -19,6 +19,7 @@
 
 #include "LightingManager.h"
 #include <app-common/zap-generated/attributes/Accessors.h>
+#include <app-common/zap-generated/cluster-objects.h>
 #include <platform/CHIPDeviceLayer.h>
 #include <platform/DeviceInstanceInfoProvider.h>
