+++ examples/thermostat/silabs/src/TemperatureManager.cpp
@@ -25,6 +25,7 @@
 #include "AppConfig.h"
 #include "AppEvent.h"
 #include "AppTask.h"
+#include <app-common/zap-generated/cluster-objects.h>
 
 /**********************************************************
  * Defines and Constants
