+++ src/app/util/BUILD.gn
@@ -54,7 +54,11 @@ source_set("types") {
 
 # This source set also depends on data-model
 source_set("af-types") {
-  sources = [ "af-types.h" ]
+  sources = [
+    "AttributesChangedListener.h",
+    "MarkAttributeDirty.h",
+    "af-types.h",
+  ]
   deps = [
     ":types",
     "${chip_root}/src/app:paths",
