+++ src/app/zap-templates/templates/app/attributes/Accessors.zapt
@@ -7,11 +7,10 @@
 
 #pragma once
 
+#include <app-common/zap-generated/cluster-enums.h>
 #include <app/data-model/Nullable.h>
-#include <app/util/af-types.h>
-#include <app/util/ember-strings.h>
-#include <app-common/zap-generated/cluster-objects.h>
-#include <lib/support/Span.h>
+#include <app/util/basic-types.h>
+#include <app/util/MarkAttributeDirty.h>
 #include <protocols/interaction_model/StatusCode.h>
 
 namespace chip {
